refactor(#258): reduce install.rs complexity — extract helpers and guidance module
- Extract is_hook_script_current() from create_hook_script, reducing
  nesting from 4 to 3 levels and isolating the idempotent version check
  as a named, testable unit
- Extract update_existing_guidance() from inject_guidance, pulling the
  5-mode existing-file branch into a dedicated helper that returns a
  bool indicating whether the caller should print the footer
- Move entire guidance section (~330 lines) into guidance.rs module,
  reducing install.rs from 1200 to ~830 lines; re-export production
  symbols through install.rs for callers and tests
- Document residual TOCTOU window in backup_settings() docstring

applies ADR-001
